Update Period class: Change duration type from uint32_t to uint16_t for improved memory efficiency and consistency in constructor parameters.

+++ firmware/Period.h
@@ -1,9 +1,9 @@
 class Period {
     public:
-        uint32_t duration = 0;
+        uint16_t duration = 0;
         uint16_t rms = 0;
 
         constexpr Period() = default;
 
-        constexpr Period(const uint32_t duration, const uint16_t rms) : duration(duration), rms(rms) {};
+        constexpr Period(const uint16_t duration, const uint16_t rms) : duration(duration), rms(rms) {};
 };
